Spring-Rest COMP-01 Tienda CART-0009: Control de excepciones NotFound implementado

+++ spring_rest/ejercicios_complementarios/tienda/src/main/java/com/spring_rest_LevelUno/tienda/controller/CarritoController.java
@@ -2,10 +2,13 @@ package com.spring_rest_LevelUno.tienda.controller;
 
 import com.spring_rest_LevelUno.tienda.entity.Carrito;
 import com.spring_rest_LevelUno.tienda.entity.Usuario;
+import com.spring_rest_LevelUno.tienda.exceptions.ResourceNotFound;
 import com.spring_rest_LevelUno.tienda.repository.CarritoRepository;
 import com.spring_rest_LevelUno.tienda.repository.DetalleRepository;
 import com.spring_rest_LevelUno.tienda.repository.UsuarioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
@@ -22,50 +25,67 @@ public class CarritoController {
    DetalleRepository detalleRepository;
 
    @GetMapping(value = "/usuarios/{id}/carritos")
-   public Iterable<Carrito> getCarritosByUsuario(@PathVariable("id") Long id) {
-      Usuario usuario = usuarioRepository.getById(id);
-      return usuario.getCarritos();
+   public ResponseEntity<?> getCarritosByUsuario(@PathVariable("id") Long id) {
+      Optional<Usuario> usuario = usuarioRepository.findById(id);
+      if (usuario.isEmpty()) {
+         throw new ResourceNotFound("¡No se encontraron carritos!");
+      }
+      return ResponseEntity.status(HttpStatus.OK).body(usuario.get().getCarritos());
    }
 
    @GetMapping(value = "/usuarios/{idUsuario}/carritos/{idCarrito}")
-   public Optional<Carrito> getCarritoById(@PathVariable("idUsuario") Long idUsuario, @PathVariable("idCarrito") Long idCarrito) {
-      Usuario usuario = usuarioRepository.getById(idUsuario);
-      return usuario.getCarritos().stream().filter(carrito -> carrito.getId().equals(idCarrito)).findFirst();
+   public ResponseEntity<?> getCarritoById(@PathVariable("idUsuario") Long idUsuario, @PathVariable("idCarrito") Long idCarrito) {
+      Optional<Usuario> usuario = usuarioRepository.findById(idUsuario);
+      Optional<Carrito> carrito = carritoRepository.findById(idCarrito);
+      if (usuario.isEmpty() || carrito.isEmpty() || !usuario.get().getCarritos().contains(carrito.get())) {
+         throw new ResourceNotFound("¡No existe el carrito solicitado!");
+      }
+      return ResponseEntity.status(HttpStatus.OK).body(carrito);
    }
 
    @PostMapping(value = "/usuarios/{id}/carritos")
-   public Carrito createCarrito(@PathVariable("id") Long id, @Valid @RequestBody Carrito requestCarrito) {
-      Usuario usuario = usuarioRepository.getById(id);
-      requestCarrito.setUsuario(usuario);
+   public ResponseEntity<?> createCarrito(@PathVariable("id") Long id, @Valid @RequestBody Carrito requestCarrito) {
+      Optional<Usuario> usuario = usuarioRepository.findById(id);
+      if (usuario.isEmpty()) {
+         throw new ResourceNotFound("¡No se pueden agregar carritos a un usuario inexistente!");
+      }
+      requestCarrito.setUsuario(usuario.get());
       requestCarrito.getDetalles().forEach(detalle -> {
          detalle.setTotal();
          detalle.setCarrito(requestCarrito);
          detalleRepository.save(detalle);
       });
-      usuario.addCarrito(requestCarrito);
-      return carritoRepository.save(requestCarrito);
+      usuario.get().addCarrito(requestCarrito);
+      return ResponseEntity.status(HttpStatus.CREATED).body(carritoRepository.save(requestCarrito));
    }
 
    @PutMapping(value = "/usuarios/{idUsuario}/carritos/{idCarrito}")
-   public Carrito updateCarrito(@PathVariable("idUsuario") Long idUsuario, @PathVariable("idCarrito") Long idCarrito, @Valid @RequestBody Carrito requestCarrito) {
-      Usuario usuario = usuarioRepository.getById(idUsuario);
-      Carrito carrito = carritoRepository.getById(idCarrito);
+   public ResponseEntity<?> updateCarrito(@PathVariable("idUsuario") Long idUsuario, @PathVariable("idCarrito") Long idCarrito, @Valid @RequestBody Carrito requestCarrito) {
+      Optional<Usuario> usuario = usuarioRepository.findById(idUsuario);
+      Optional<Carrito> carrito = carritoRepository.findById(idCarrito);
+      if (usuario.isEmpty() || carrito.isEmpty() || !usuario.get().getCarritos().contains(carrito.get())) {
+         throw new ResourceNotFound("¡No se encontró el carrito solicitado!");
+      }
       requestCarrito.getDetalles().forEach(detalle -> {
-         detalle.setCarrito(carrito);
+         detalle.setCarrito(carrito.get());
          detalle.setTotal();
          if (detalle.getCantidad().equals(0)) {
             requestCarrito.getDetalles().remove(detalle);
          }
       });
-      carrito.setDetalles(requestCarrito.getDetalles());
-      return carritoRepository.saveAndFlush(carrito);
+      carrito.get().setDetalles(requestCarrito.getDetalles());
+      return ResponseEntity.status(HttpStatus.OK).body(carritoRepository.saveAndFlush(carrito.get()));
    }
 
    @DeleteMapping(value = "/usuarios/{idUsuario}/carritos/{idCarrito}")
-   public void deleteCarrito(@PathVariable("idUsuario") Long idUsuario, @PathVariable("idCarrito") Long idCarrito) {
-      Usuario usuario = usuarioRepository.getById(idUsuario);
-      Carrito carrito = carritoRepository.getById(idCarrito);
-      usuario.getCarritos().remove(carrito);
+   public ResponseEntity<?> deleteCarrito(@PathVariable("idUsuario") Long idUsuario, @PathVariable("idCarrito") Long idCarrito) {
+      Optional<Usuario> usuario = usuarioRepository.findById(idUsuario);
+      Optional<Carrito> carrito = carritoRepository.findById(idCarrito);
+      if (usuario.isEmpty() || carrito.isEmpty() || !usuario.get().getCarritos().contains(carrito.get())) {
+         throw new ResourceNotFound("¡El carrito que desea eliminar no existe!");
+      }
+      usuario.get().getCarritos().remove(carrito.get());
       carritoRepository.deleteById(idCarrito);
+      return ResponseEntity.status(HttpStatus.OK).body("Eliminado");
    }
 }
